Automation: move job tests to a new package
