+++ addOns/automation/src/main/java/org/zaproxy/addon/automation/gui/AddTestDialog.java
@@ -23,13 +23,13 @@ import java.util.ArrayList;
 import java.util.List;
 import org.parosproxy.paros.Constant;
 import org.parosproxy.paros.view.View;
-import org.zaproxy.addon.automation.AbstractAutomationTest;
-import org.zaproxy.addon.automation.AutomationAlertTest;
 import org.zaproxy.addon.automation.AutomationEventPublisher;
 import org.zaproxy.addon.automation.AutomationJob;
-import org.zaproxy.addon.automation.AutomationStatisticTest;
 import org.zaproxy.addon.automation.jobs.ActiveScanJob;
 import org.zaproxy.addon.automation.jobs.PassiveScanWaitJob;
+import org.zaproxy.addon.automation.tests.AbstractAutomationTest;
+import org.zaproxy.addon.automation.tests.AutomationAlertTest;
+import org.zaproxy.addon.automation.tests.AutomationStatisticTest;
 import org.zaproxy.zap.utils.DisplayUtils;
 import org.zaproxy.zap.view.StandardFieldsDialog;
