+++ addOns/automation/src/main/java/org/zaproxy/addon/automation/tests/AutomationStatisticTest.java
@@ -17,7 +17,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.zaproxy.addon.automation;
+package org.zaproxy.addon.automation.tests;
 
 import java.util.Arrays;
 import java.util.LinkedHashMap;
@@ -25,6 +25,8 @@ import java.util.Map;
 import org.apache.commons.lang.StringUtils;
 import org.parosproxy.paros.Constant;
 import org.parosproxy.paros.control.Control;
+import org.zaproxy.addon.automation.AutomationJob;
+import org.zaproxy.addon.automation.AutomationProgress;
 import org.zaproxy.addon.automation.gui.StatisticTestDialog;
 import org.zaproxy.addon.automation.jobs.JobUtils;
 import org.zaproxy.zap.extension.stats.ExtensionStats;
