+++ addOns/retest/retest.gradle.kts
@@ -10,7 +10,7 @@ zapAddOn {
         dependencies {
             addOns {
                 register("automation") {
-                    version.set(">=0.5.0")
+                    version.set(">=0.6.0")
                 }
             }
         }
